started working on profile set up
