+++ src/components/SetUpProfile/SetUpProfileList.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import SetUpProfileItem from './SetUpProfileItem';
+
+const SetUpProfileList = ({
+  authUser,
+  biodata,
+}) => (
+  <ul>
+    {biodata.map(biodata => (
+      <MessageItem
+        authUser={authUser}
+        key={biodata.uid}
+        biodata={biodata}
+      />
+    ))}
+  </ul>
+);
+
+export default SetUpProfileList;
